corrected the key's data type of each repo

+++ src/main/java/com/depot/ims/repositories/AuditsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
-public interface AuditsRepository extends JpaRepository<Audit, Integer> {
+public interface AuditsRepository extends JpaRepository<Audit, Long> {
     List<Audit> findAll();
 
     @Query("SELECT a FROM Audit a WHERE a.auditId=?1")
